requirements.txt
Django
psycopg2-binary
djangorestframework
# markdown
django-filter
djangorestframework-jwt
django-cors-headers
gunicorn

boto3
django-storages